+++ CHANGELOG.md
@@ -4,3 +4,4 @@
 - Typos
 - Improve idtrackerai/network file structure and imports
 - Improve v4 compatibility reading video.json/npy
+- NetworkParams as dataclass